EZP-27428: Split MultiLanguaveValueTrait into separate traits
